// src/lib/types/collections.ts
export interface UserDataDocumentsType extends Models.Row {
  /**
   * The avatar ID of the user.
   */
  avatarId: string | null
  /**
   * The banner ID of the user.
   */
  profileBannerId: string | null
  /**
   * The user's status.
   */
  status: string | null
  /**
   * The user's display name.
   */
  displayName: string
  /**
   * The user's bio.
   */
  bio: string | null
  /**
   * The user's birthday.
   * @example '2000-01-01T00:00:00:000Z'
   */
  birthday: string | null
  /**
   * The user's profile URL.
   */
  profileUrl: string
  /**
   * The user's pronouns.
   */
  pronouns: string | null
  /**
   * The user's discord username. (Not the ID)
   */
  discordname: string | null
  /**
   * The user's telegram name.
   */
  telegramname: string | null
  /**
   * The user's furaffinity name.
   */
  furaffinityname: string | null
  /**
   * The user's x/twitter name.
   */
  X_name: string | null
  /**
   * The user's twitch name.
   */
  twitchname: string | null
  /**
   * The user's bluesky name.
   */
  blueskyname: string | null
  /**
   * The user's location.
   */
  location: string | null
  /**
   * The user's badges.
   */
  badges: string[]
  /**
   * The user's followers count.
   */
  followersCount: number
  /**
   * The user's following count.
   */
  followingCount: number
}
// ...
export interface CommunityDocumentsType extends Models.Row {
  /**
   * The name of the community.
   */
  name?: string | null
  /**
   * The description of the community.
   */
  description: string
  /**
   * The tags of the community for searching.
   */
  tags: string[]
  /**
   * The banner ID of the community
   */
  bannerId: string
  /**
   * The avatar ID of the community.
   */
  avatarId: string
  /**
   * The status of the community.
   */
  status: string
  /**
   * The community settings.
   */
  communitySettings: CommunitySettingsDocumentsType
  /**
   * The amount of followers the community has.
   */
  followersCount: number
  /**
   * If user is following the community.
   */
  isFollowing: boolean
  /**
   * The community preferences.
   */
  prefs?: CommunityPrefsDocumentsType
  /**
   * The roles of the community.
   */
  roles?: string[]
}

export interface CommunityPrefsType {
  total: number
  rows: CommunityPrefsDocumentsType[]
}

export interface CommunityPrefsDocumentsType extends Models.Row {
  /**
   * The community ID.
   */
  communityId: string
  /**
   * If the community is blocked.
   */
  isBlocked: boolean | null
  /**
   * If the community is muted.
   */
  isMuted: boolean | null
}

export interface CommunitySettingsType {
  total: number
  rows: CommunitySettingsDocumentsType[]
}

export interface CommunitySettingsDocumentsType extends Models.Row {
  /**
   * If the community is indexed by search engines and findable.
   */
  isFindable: boolean
  /**
   * If the community is publicly accessible.
   */
  hasPublicPage: boolean
  /**
   * If the community is nsfw.
   */
  nsfw: boolean
}